requirements.txt
streamlit==1.24.0 
openai==1.51.2 
python-dotenv==1.0.0 
